requirements.txt
langchain
langchain_google_genai
python-dotenv
langchain-core
